Improved average feature coefficient calculation in notebook
Previous method to calculate average feature coefficient and error bars was incorrect, so switched to using 10 random splits for training instead of regular k-fold cross validation which was not shuffling records

Scrapped old confidence interval calculation and switched to just standard error.

Re-reviewed for categories, and updated Results.xlsx. This has been imported, but visualizations need to be reviewed and updated in ReadMe and presentation, and need to check for new or changed categories from initial results.