New visualizations with standard error bars
No change in recommendations, but updated visualizations in presentation and README file. A few subjects and n-grams were different after running 10 random splits of the data versus 5-fold cross validation, but they did not affect the makeup of categories used in the recommendations. I hadn't thought about the fact the regular cross validation does not shuffle the splits, and since they were arranged chronologically, some topics and themes tended to be heavier in earlier or later years. Randomly shuffled splits addressed this.